fixed signaling with block tweaks
